requirements.txt
fastapi
uvicorn
httpx
python-multipart
python-dotenv
fastapi
uvicorn
